backend/requirements.txt
Flask
Flask-Cors
joblib
scikit-learn
pandas
datasets
pymongo
flask-bcrypt
flask-jwt-extended
python-dotenv
google-generativeai